Filter common markers between strata

The function will filter the markers by keeping only those in common between all strata (population or any groupings defined in STRATA column).

Filter targets: SNPs

Statistics: strata genotyping rate per SNPs

Used internally in radiator and might be of interest for users who wants to keep only markers in common.

filter_common_markers(
  data,
  filter.common.markers = TRUE,
  fig = FALSE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)

Arguments

data

(4 options) A file or object generated by radiator:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Look into tidy_genomic_data, read_vcf or tidy_vcf.

filter.common.markers

(optional, logical) Default: filter.common.markers = TRUE.

fig

(optional, logical) fig = TRUE will produce an UpSet fig to visualize the number of markers between populations. The package is required for this to work... Default: fig = FALSE.

parallel.core

(optional) The number of core used for parallel execution during import. Default: parallel.core = parallel::detectCores() - 1.

verbose

(optional, logical) verbose = TRUE to be chatty during execution. Default: verbose = FALSE.

...

(optional) To pass further arguments for fine-tuning the function and legacy arguments.

Value

A list with the filtered input, whitelist and blacklist of markers..

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) {
require(SeqArray) # when using gds
common <- radiator::filter_common_markers(data = "my.radiator.gds.rad", verbose = TRUE)
}