Fis filter

TODO

filter_fis(
  data,
  approach = "haplotype",
  fis.min.threshold,
  fis.max.threshold,
  fis.diff.threshold,
  pop.threshold,
  percent,
  filename
)

Arguments

data

14 options for input (diploid data only): VCFs (SNPs or Haplotypes, to make the vcf population ready), plink (tped, bed), stacks haplotype file, genind (library(adegenet)), genlight (library(adegenet)), gtypes (library(strataG)), genepop, DArT, and a data frame in long/tidy or wide format. To verify that radiator detect your file format use detect_genomic_format (see example below). Documented in Input genomic datasets of tidy_genomic_data.

DArT and VCF data: radiator was not meant to generate alleles and genotypes if you are using a VCF file with no genotype (only genotype likelihood: GL or PL). Neither is radiator able to magically generate a genind object from a SilicoDArT dataset. Please look at the first few lines of your dataset to understand it's limit before asking raditor to convert or filter your dataset.

approach

Character. By "SNP" or by "haplotype". The function will consider the SNP or haplotype statistics to filter the marker. Default by "haplotype".

fis.min.threshold

Number.

fis.max.threshold

Number.

fis.diff.threshold

Number (0 - 1)

pop.threshold

Fixed number of pop required to keep the locus.

percent

Is the threshold a percentage ? TRUE or FALSE.

filename

(optional) The function uses write.fst, to write the tidy data frame in the folder created in the working directory. The file extension appended to the filename provided is .rad. Default: filename = NULL.

Details

The Fis calculated uses the ratio of averages (1-mean(Ho)/mean(Hs)) and NOT THE AVERAGE OF RATIOS (Nei 1987).

References

Nei M. (1987) Molecular Evolutionary Genetics. Columbia University Press.

Author

Thierry Gosselin thierrygosselin@icloud.com