Filter dataset with whitelist of markers

Filter dataset with whitelist of markers Used internally in radiator and might be of interest for users.

filter_whitelist(data, whitelist.markers = NULL, verbose = TRUE, ...)

Arguments

data
(4 options) A file or object generated by radiator:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Look into tidy_genomic_data, read_vcf or tidy_vcf.

whitelist.markers
(path or object) The whitelist is an object in your global environment or a file in the working directory (e.g. "whitelist.txt"). The dataframe contains one, a combination or all of these columns: MARKERS, CHROM, LOCUS, POS. Columns are cleaned of separators that interfere with some packages or codes, detailed in clean_markers_names. Default whitelist.markers = NULL.

verbose
(optional, logical) When verbose = TRUE the function is a little more chatty during execution. Default: verbose = TRUE.

...
(optional) Advance mode that allows to pass further arguments for fine-tuning the function. Also used for legacy arguments (see details or special section)

See also

Examples

# NOT RUN {
data <- radiator::filter_whitelist(data = data, whitelist.markers = "mywhitelist.tsv")
# }