radiator common arguments

radiator common arguments

radiator_common_arguments(
  interactive.filter = TRUE,
  gds,
  data,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  random.seed = NULL,
  ...
)

Arguments

interactive.filter

(optional, logical) Do you want the filtering session to be interactive. Figures of distribution are shown before asking for filtering thresholds. Default: interactive.filter = TRUE.

gds

(2 options) A Genomic Data Structure (GDS) file or object

How to get GDS ? Look into tidy_genomic_data, read_vcf, tidy_vcf or write_gds.

data

(4 options) A file or object generated by radiator:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Look into tidy_genomic_data, read_vcf or tidy_vcf.

parallel.core

(optional) The number of core used for parallel execution during import. Default: parallel.core = parallel::detectCores() - 1.

verbose

(optional, logical) When verbose = TRUE the function is a little more chatty during execution. Default: verbose = TRUE.

random.seed

(integer, optional) For reproducibility, set an integer that will be used inside the function that requires randomness. With default, a random number is generated and printed in the appropriate output. Default: random.seed = NULL.

...

(optional) Advance mode that allows to pass further arguments for fine-tuning the function. Also used for legacy arguments (see details or special section)