read whitelist of markers

Read a whitelist object or file.

Used internally in radiator and might be of interest for users.

read_whitelist(whitelist.markers = NULL, verbose = FALSE)

Arguments

whitelist.markers
(path or object) The whitelist is an object in your global environment or a file in the working directory (e.g. "whitelist.txt"). The dataframe contains one, a combination or all of these columns: MARKERS, CHROM, LOCUS, POS. Columns are cleaned of separators that interfere with some packages or codes, detailed in clean_markers_names. Default whitelist.markers = NULL.

verbose
(optional, logical) When verbose = TRUE the function is a little more chatty during execution. Default: verbose = TRUE.

Details

multi allelic datasets: Example: VCF with haplotypes containing CHROM, LOCUS, POS columns: If the whitelist was not created from the same dataset, the filtering could result in losing all the markers. The POS column is different in biallelic and multiallelic file...

See also

Examples

if (FALSE) {
wl <- radiator::read_whitelist(data = data, whitelist.markers = "mywhitelist.tsv")
}