Tidy a gtypes object to a tidy dataframe

Transform a [strataG gtypes](https://github.com/EricArcher/strataG) object from to a tidy dataframe. Used internally in radiator and might be of interest for users.

tidy_gtypes(data)

Arguments

data
A gtypes object (>= v.2.0.2) in the global environment.

References

Archer FI, Adams PE, Schneiders BB. strataG: An r package for manipulating, summarizing and analysing population genetic data. Molecular Ecology Resources. 2017; 17: 5-11. doi:10.1111/1755-0998.12559

Author

Thierry Gosselin thierrygosselin@icloud.com