Write an arlequin file from a tidy data frame

Write a arlequin file from a tidy data frame. Used internally in radiator and assigner and might be of interest for users.

write_arlequin(data, pop.levels = NULL, filename = NULL, ...)

Arguments

data
A tidy data frame object in the global environment or a tidy data frame in wide or long format in the working directory. How to get a tidy data frame ? Look into radiator tidy_genomic_data.

pop.levels
(optional, string) A character string with your populations ordered. Default: pop.levels = NULL.

filename
(optional) The file name prefix for the arlequin file written to the working directory. With default: filename = NULL, the date and time is appended to radiator_arlequin_.

...
other parameters passed to the function.

Value

An arlequin file is saved to the working directory.

References

Excoffier, L.G. Laval, and S. Schneider (2005) Arlequin ver. 3.0: An integrated software package for population genetics data analysis. Evolutionary Bioinformatics Online 1:47-50.