Write a betadiv file from a tidy data frame

Write a betadiv file from a tidy data frame. Used internally in radiator and assigner and might be of interest for users.

write_betadiv(data)

Arguments

data
A tidy data frame object in the global environment or a tidy data frame in wide or long format in the working directory. How to get a tidy data frame ? Look into radiator tidy_genomic_data.

Value

A betadiv object is returned.

References

Lamy T, Legendre P, Chancerelle Y, Siu G, Claudet J (2015) Understanding the Spatio-Temporal Response of Coral Reef Fish Communities to Natural Disturbances: Insights from Beta-Diversity Decomposition. PLoS ONE, 10, e0138696.

See also

Author

Laura Benestan laura.benestan@icloud.com and Thierry Gosselin thierrygosselin@icloud.com