Detect reference genome

Detect if the dataset was de novo or reference genome assembled

detect_ref_genome(chromosome = NULL, data = NULL, verbose = TRUE)

Arguments

chromosome

(string) String with chromosome ID, unique or not. Default: chromosome = NULL.

data

(optional, path or object) radiator GDS object or file. When this argument is used, the function look inside the GDS radiator node for:

reference.genome node and TRUE/FALSE value

markers.meta node and detect the information using the CHROM column

chromosome node directly.

Default: data = NULL.

verbose

(optional, logical) When verbose = TRUE the function is a little more chatty during execution. Default: verbose = TRUE.

Value

TRUE/FALSE if dataset is assembled with a reference genome or not.

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) { # \dontrun{
ref.genome <- radiator::detect_ref_genome(chromosome = tidy.data$CHROM)
} # }