Filter monomorphic markers

Filter monomorphic markers. This filter will remove from the dataset markers with just one genotype phenotype:

genotypes are ALL homozygotes REF/REF (pp)

genotypes are ALL heterozygotes REF/ALT, ALT/REF (pq or qp)

genotypes are ALL homozygotes ALT/ALT (qq)

Filter targets: SNPs

Statistics: the number of genotype phenotypes

Used internally in radiator and might be of interest for users who wants to keep only polymorphic markers in their dataset.

filter_monomorphic(
  data,
  filter.monomorphic = TRUE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)

Arguments

data
(4 options) A file or object generated by radiator:

tidy data

Genomic Data Structure (GDS)

How to get GDS and tidy data ? Look into tidy_genomic_data, read_vcf or tidy_vcf.

filter.monomorphic
(optional, logical) Default: filter.monomorphic = TRUE.

parallel.core
(optional) The number of core used for parallel execution during import. Default: parallel.core = parallel::detectCores() - 1.

verbose
(optional, logical) When verbose = TRUE the function is a little more chatty during execution. Default: verbose = TRUE.

...
(optional) Advance mode that allows to pass further arguments for fine-tuning the function. Also used for legacy arguments (see details or special section)

Value

A list with the filtered input, whitelist and blacklist of markers..

See also

Author

Thierry Gosselin thierrygosselin@icloud.com

Examples

if (FALSE) {
require(SeqArray) # when using gds
mono <- radiator::filter_monomorphic(data = "my.radiator.gds.rad", verbose = TRUE)
}