Write a BayeScan file from a tidy data frame

Write a BayeScan file from a tidy data frame. The data is bi- or multi-allelic. Used internally in radiator and might be of interest for users.

write_bayescan(data, pop.select = NULL, filename = NULL,
  parallel.core = parallel::detectCores() - 1, ...)

Arguments

data
A tidy data frame object in the global environment or a tidy data frame in wide or long format in the working directory. How to get a tidy data frame ? Look into radiator tidy_genomic_data.

pop.select
(optional, string) Selected list of populations for the analysis. e.g. pop.select = c("QUE", "ONT") to select QUE and ONT population samples (out of 20 pops). Default: pop.select = NULL

filename
(optional) The file name prefix for the bayescan file written to the working directory. With default: filename = NULL, the date and time is appended to radiator_bayescan_.

parallel.core
(optional) The number of core used for parallel execution during import. Default: parallel.core = parallel::detectCores() - 1.

...
(optional) To pass further arguments for fine-tuning the function.

Value

A bayescan file is written in the working directory.

Details

Integrated filters:

by defaults only markers found in common between populations are used

by defaults monomorphic markers are automatically removed before generating the dataset.

References

Foll, M and OE Gaggiotti (2008) A genome scan method to identify selected loci appropriate for both dominant and codominant markers: A Bayesian perspective. Genetics 180: 977-993

Foll M, Fischer MC, Heckel G and L Excoffier (2010) Estimating population structure from AFLP amplification intensity. Molecular Ecology 19: 4638-4647

Fischer MC, Foll M, Excoffier L and G Heckel (2011) Enhanced AFLP genome scans detect local adaptation in high-altitude populations of a small rodent (Microtus arvalis). Molecular Ecology 20: 1450-1462