Write a related file from a tidy data frame

Write a related file from a tidy data frame. This output file format enables to run the data in the related R package (Pew et al. 2015), which is essantially the R version of COANCESTRY fortran program developed by Jinliang Wang. Used internally in radiator and might be of interest for users.

write_related(
  data,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  ...
)

Arguments

data

A tidy data frame object in the global environment or a tidy data frame in wide or long format in the working directory. How to get a tidy data frame ? Look into radiator tidy_genomic_data.

filename

(optional) The file name prefix for the related file written to the working directory. With default: filename = NULL, the date and time is appended to radiator_related_.

parallel.core

(optional) The number of core used for parallel execution during import. Default: parallel.core = parallel::detectCores() - 1.

...

other parameters passed to the function.

Value

A related file is saved to the working directory.

References

Pew J, Muir PH, Wang J, Frasier TR (2015) related: an R package for analysing pairwise relatedness from codominant molecular markers. Molecular Ecology Resources, 15, 557-561.

Wang, J. 2011. COANCESTRY: A program for simulating, estimating and analysing relatedness and inbreeding coefficients. Molecular Ecology Resources 11(1): 141-145.

Author

Thierry Gosselin thierrygosselin@icloud.com